requirements.txt
alabaster==0.7.10
argh==0.26.2
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
CommonMark==0.5.4
coverage==4.5.1
decorator==4.2.1
dj-database-url==0.5.0
Django==1.11.3
django-cors-headers==2.1.0
django-extensions==1.8.1
-e git+https://github.com/fernandoe/django-fe-core.git@release/0.4.0#egg=django_fe_core
-e git+https://github.com/fernandoe/django-fe-jwt.git@release/0.0.1#egg=django_fe_jwt
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
docutils==0.14
drf-nested-routers==0.90.0
factory-boy==2.8.1
Faker==0.8.12
idna==2.5
imagesize==1.0.0
ipaddress==1.0.18
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
livereload==2.5.1
MarkupSafe==1.0
mysqlclient==1.3.12
oauthlib==2.0.7
parso==0.1.1
pathtools==0.1.2
pexpect==4.4.0
pickleshare==0.7.4
port-for==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
Pygments==2.2.0
PyJWT==1.6.1
pytest==3.1.3
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-runner==2.11.1
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
recommonmark==0.4.0
requests==2.18.1
requests-mock==1.3.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
text-unidecode==1.2
tornado==5.0.1
traitlets==4.3.2
urllib3==1.21.1
watchdog==0.8.3
wcwidth==0.1.7
